refactor : 팝업 detail, list 합치기
