+++ src/main/java/org/example/popspace/controller/popup/PopupController.java
@@ -1,5 +1,6 @@
 package org.example.popspace.controller.popup;
 
+import java.text.ParseException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -8,8 +9,9 @@ import org.example.popspace.dto.popup.LikeResponseDto;
 import org.example.popspace.dto.popup.LikeUpdateRequestDto;
 import org.example.popspace.dto.popup.PopupDetailResponseDto;
 import org.example.popspace.dto.popup.PopupInfoDto;
+import org.example.popspace.dto.popup.PopupListDto;
 import org.example.popspace.dto.popup.ReviewDto;
-import org.example.popspace.service.popup.PopupDetailService;
+import org.example.popspace.service.popup.PopupService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,20 +19,21 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
 @RestController
-@RequestMapping("/popup/detail")
+@RequestMapping("/popup")
 @RequiredArgsConstructor
-public class PopupDetailController {
-    private final PopupDetailService popupDetailService;
+public class PopupController {
+    private final PopupService popupService;
 
     /* 팝업 상세 페이지 조회 (상세, 리뷰) */
-    @GetMapping("/info-review/{popupId}")
+    @GetMapping("/detail/{popupId}")
     public ResponseEntity<PopupDetailResponseDto> infoAndReview(@PathVariable Long popupId) {
-        PopupInfoDto popupInfo = popupDetailService.findPopupInfoByPopupId(popupId);
-        List<ReviewDto> reviewDtoList = popupDetailService.findReviewByPopupId(popupId);
+        PopupInfoDto popupInfo = popupService.findPopupInfoByPopupId(popupId);
+        List<ReviewDto> reviewDtoList = popupService.findReviewByPopupId(popupId);
 
         PopupDetailResponseDto responseDto = PopupDetailResponseDto.builder()
                 .popupInfo(popupInfo)
@@ -41,10 +44,10 @@ public class PopupDetailController {
     }
 
     /* 팝업 상세 페이지 조회 (찜) */
-    @GetMapping("/reserve-like/{popupId}")
+    @GetMapping("/like/{popupId}")
     public ResponseEntity<LikeResponseDto> reserveAndLike(@PathVariable Long popupId,
                                                           @AuthenticationPrincipal CustomUserDetail userDetail) {
-        String isPopupLike = popupDetailService.findPopupLikeByPopupIdMemberId(popupId, userDetail.getId());
+        String isPopupLike = popupService.findPopupLikeByPopupIdMemberId(popupId, userDetail.getId());
         boolean isLiked = "ACTIVE".equals(isPopupLike); // 찜 상태 확인
 
         LikeResponseDto likeResponseDto = LikeResponseDto.builder()
@@ -63,8 +66,8 @@ public class PopupDetailController {
                                                                @AuthenticationPrincipal CustomUserDetail userDetail) {
         log.info("popup/detail/like-update: popupId={}, memberId={}, toBeState={}", dto.getPopupId(), userDetail.getId(), dto.isToBeState());
 
-        popupDetailService.updatePopupLike(dto.getPopupId(), userDetail.getId(), dto.isToBeState());
-        String isPopupLike = popupDetailService.findPopupLikeByPopupIdMemberId(dto.getPopupId(), userDetail.getId()); // 찜 상태 확인
+        popupService.updatePopupLike(dto.getPopupId(), userDetail.getId(), dto.isToBeState());
+        String isPopupLike = popupService.findPopupLikeByPopupIdMemberId(dto.getPopupId(), userDetail.getId()); // 찜 상태 확인
         log.info("업데이트된 찜 상태: {}", isPopupLike);
 
         boolean isLiked = "ACTIVE".equals(isPopupLike); // 찜 상태 확인
@@ -75,4 +78,19 @@ public class PopupDetailController {
         return ResponseEntity.ok(likeResponseDto);
     }
 
+    /* 팝업 목록 */
+    @GetMapping("/list")
+    public PopupListDto getPopupList(@AuthenticationPrincipal CustomUserDetail userDetail,
+                                     @RequestParam(required = false) String searchKeyword,
+                                     @RequestParam (required = false) String searchDate,
+                                     @RequestParam (required = false) String sortKey) throws ParseException {
+
+        log.info("/popup/list : searchKeyword={}, searchDate={}, sortKey={}", searchKeyword, searchDate, sortKey);
+        PopupListDto popupListDto = PopupListDto.builder()
+                .popupList(popupService.getPopupList(userDetail.getId(), searchKeyword, searchDate, sortKey))
+                .build();
+
+        log.info("조회된 팝업 개수: {}", popupListDto.getPopupList().size());
+        return popupListDto;
+    }
 }
